Extend origin handling to structural changes

+++ src/Umbraco.Cms.Search.Core/NotificationHandlers/RebuildIndexesNotificationHandler.cs
@@ -2,19 +2,23 @@
 using Microsoft.Extensions.Options;
 using Umbraco.Cms.Core.Events;
 using Umbraco.Cms.Core.Models;
-using Umbraco.Cms.Core.Notifications;
+using Umbraco.Cms.Core.Scoping;
 using Umbraco.Cms.Core.Services.Changes;
+using Umbraco.Cms.Search.Core.Cache.ContentType;
+using Umbraco.Cms.Search.Core.Cache.Language;
+using Umbraco.Cms.Search.Core.Cache.MediaType;
+using Umbraco.Cms.Search.Core.Cache.MemberType;
 using Umbraco.Cms.Search.Core.Configuration;
 using Umbraco.Cms.Search.Core.Models.Configuration;
 using Umbraco.Cms.Search.Core.Services.ContentIndexing;
 
 namespace Umbraco.Cms.Search.Core.NotificationHandlers;
 
-internal sealed class RebuildIndexesNotificationHandler :
-    INotificationAsyncHandler<LanguageDeletedNotification>,
-    INotificationAsyncHandler<ContentTypeChangedNotification>,
-    INotificationAsyncHandler<MemberTypeChangedNotification>,
-    INotificationAsyncHandler<MediaTypeChangedNotification>
+internal sealed class RebuildIndexesNotificationHandler  : IndexingNotificationHandlerBase,
+    INotificationAsyncHandler<LanguageCacheRefresherNotification>,
+    INotificationAsyncHandler<ContentTypeCacheRefresherNotification>,
+    INotificationAsyncHandler<MemberTypeCacheRefresherNotification>,
+    INotificationAsyncHandler<MediaTypeCacheRefresherNotification>
 {
     private readonly IContentIndexingService _contentIndexingService;
     private readonly IIndexDocumentService _indexDocumentService;
@@ -25,7 +29,9 @@ internal sealed class RebuildIndexesNotificationHandler :
         IContentIndexingService contentIndexingService,
         IIndexDocumentService indexDocumentService,
         ILogger<RebuildIndexesNotificationHandler> logger,
-        IOptions<IndexOptions> options)
+        IOptions<IndexOptions> options,
+        ICoreScopeProvider coreScopeProvider)
+        : base(coreScopeProvider)
     {
         _contentIndexingService = contentIndexingService;
         _indexDocumentService = indexDocumentService;
@@ -33,33 +39,50 @@ internal sealed class RebuildIndexesNotificationHandler :
         _options = options.Value;
     }
 
-    public async Task HandleAsync(LanguageDeletedNotification notification, CancellationToken cancellationToken)
+    public async Task HandleAsync(LanguageCacheRefresherNotification notification, CancellationToken cancellationToken)
     {
-        _logger.LogInformation("Rebuilding search indexes after language deletion...");
+        LanguageCacheRefresher.JsonPayload[] payloads = GetNotificationPayloads<LanguageCacheRefresher.JsonPayload>(notification, out var origin);
+        if (payloads.Any(payload => payload.ChangeTypes is LanguageChangeTypes.Delete) is false)
+        {
+            return;
+        }
+
         await _indexDocumentService.DeleteAllAsync();
 
         foreach (ContentIndexRegistration indexRegistration in _options.GetContentIndexRegistrations())
         {
             if (indexRegistration.ContainedObjectTypes.Contains(UmbracoObjectTypes.Document))
             {
-                _contentIndexingService.Rebuild(indexRegistration.IndexAlias);
+                _contentIndexingService.Rebuild(indexRegistration.IndexAlias, origin);
             }
         }
     }
 
-    public async Task HandleAsync(ContentTypeChangedNotification notification, CancellationToken cancellationToken)
-        => await RebuildByObjectType(notification.Changes, UmbracoObjectTypes.Document);
-
-    public async Task HandleAsync(MemberTypeChangedNotification notification, CancellationToken cancellationToken)
-        => await RebuildByObjectType(notification.Changes, UmbracoObjectTypes.Member);
-
-    public async Task HandleAsync(MediaTypeChangedNotification notification, CancellationToken cancellationToken)
-        => await RebuildByObjectType(notification.Changes, UmbracoObjectTypes.Media);
-
-    private async Task RebuildByObjectType<T>(IEnumerable<ContentTypeChange<T>> changes, UmbracoObjectTypes objectType)
-        where T : class, IContentTypeComposition
+    public async Task HandleAsync(ContentTypeCacheRefresherNotification notification, CancellationToken cancellationToken)
     {
-        foreach (ContentTypeChange<T> change in changes)
+        ContentTypeCacheRefresher.JsonPayload[] payloads = GetNotificationPayloads<ContentTypeCacheRefresher.JsonPayload>(notification, out var origin);
+
+        await HandleContentTypeChangesAsync(payloads.Select(payload => (payload.ContentTypeKey, payload.ChangeTypes)), UmbracoObjectTypes.Document, origin);
+    }
+
+    public async Task HandleAsync(MemberTypeCacheRefresherNotification notification, CancellationToken cancellationToken)
+    {
+        MemberTypeCacheRefresher.JsonPayload[] payloads = GetNotificationPayloads<MemberTypeCacheRefresher.JsonPayload>(notification, out var origin);
+
+        await HandleContentTypeChangesAsync(payloads.Select(payload => (payload.MemberTypeKey, payload.ChangeTypes)), UmbracoObjectTypes.Member, origin);
+    }
+
+    public async Task HandleAsync(MediaTypeCacheRefresherNotification notification, CancellationToken cancellationToken)
+    {
+        MediaTypeCacheRefresher.JsonPayload[] payloads = GetNotificationPayloads<MediaTypeCacheRefresher.JsonPayload>(notification, out var origin);
+
+        await HandleContentTypeChangesAsync(payloads.Select(payload => (payload.MediaTypeKey, payload.ChangeTypes)), UmbracoObjectTypes.Media, origin);
+    }
+
+    private async Task HandleContentTypeChangesAsync(IEnumerable<(Guid ContentTypeKey, ContentTypeChangeTypes ChangeTypes)> changes, UmbracoObjectTypes objectType, string origin)
+    {
+        // TODO: rewrite this logic, it will potentially rebuild multiple times
+        foreach ((Guid ContentTypeKey, ContentTypeChangeTypes ChangeTypes) change in changes)
         {
             if (change.ChangeTypes is not (ContentTypeChangeTypes.RefreshMain or ContentTypeChangeTypes.Remove))
             {
@@ -71,7 +94,7 @@ internal sealed class RebuildIndexesNotificationHandler :
                 if (indexRegistration.ContainedObjectTypes.Contains(objectType))
                 {
                     await _indexDocumentService.DeleteAllAsync();
-                    _contentIndexingService.Rebuild(indexRegistration.IndexAlias);
+                    _contentIndexingService.Rebuild(indexRegistration.IndexAlias, origin);
                 }
             }
         }
